Dice game
Rules:

When the game starts, player 1 clicks on the button "Roll dice". Each click numbers on dices sums up and player can see his points in "Current" and it can be countinued until player gets dice with 1, because when player gets dice 1, the game for him stops and player 2 starts to click. So that we have the button "Hold". If after some clicks player feel that he is enough for his turn, player can click on "Hold" so that his points saved and turn come to another player. You can play until someone collects 100 points and game ends.